% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunsForOptimalV2.R
\name{HessFun}
\alias{HessFun}
\title{HessFun}
\usage{
HessFun(bb, ixx, iyy, iw, iZZ)
}
\arguments{
\item{bb}{initial values for the intercept and slope coefficients}

\item{ixx}{continuous predictor}

\item{iyy}{binary outcome}

\item{iw}{the weighted parameter}

\item{iZZ}{covariates to be incorporated in the model}
}
\value{
the Hessian matrix
}
\description{
calculate the Hessian matrix
}
\examples{
iZZ <- matrix(0,100,10)
ixx <- matrix(0,100,1)
iyy <- matrix(0,100,1)
iw <-matrix(0,100,1)
# HessFun <- function(bb, ixx, iyy, iw, iZZ)

}
