% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProLikFun.r
\name{myfit.prof}
\alias{myfit.prof}
\title{myfit.prof}
\usage{
myfit.prof(Predictor, Outcome)
}
\arguments{
\item{Predictor}{predictor in the model}

\item{Outcome}{response in the model}
}
\value{
beta0, beta1, lambda, iteration

calculate the beta0, beta1, lambda, iteration in LBC model
}
\description{
this function is aiming at estimates the LBC model parameter
}
\examples{
\donttest{xx <- runif(100,min=0,max=1)}
\donttest{yy <- rbind(matrix(0,50,1),matrix(1,50,1))}
\donttest{myprofit <- myfit.prof(xx, yy)}
\donttest{init <- myprofit[1:3]}
#mymle <- maxLik(logLik=LogLikeFun, grad = ScoreFun, start=init, ixx=xx, iyy=yy )
#beta0 <- mymle$estimate[1]
#beta1 <- mymle$estimate[2]
#lambda <- mymle$estimate[3]
#yy.gen <- beta0 + beta1*(xx^lambda-1)/lambda
}
